Added security mechanisms for docker commands
