+++ docker/src/main/java/edu/kit/kastel/informalin/framework/docker/DockerContainer.java
@@ -1,0 +1,8 @@
+/* Licensed under MIT 2022. */
+package edu.kit.kastel.informalin.framework.docker;
+
+public record DockerContainer(String id, String image, String status, String name) {
+    public boolean isRunning() {
+        return status().startsWith("Up");
+    }
+}
