+++ docker/src/main/java/edu/kit/kastel/informalin/framework/docker/DockerImage.java
@@ -1,0 +1,14 @@
+/* Licensed under MIT 2022. */
+package edu.kit.kastel.informalin.framework.docker;
+
+import java.util.List;
+
+public record DockerImage(String repository, String tag, List<Integer> exposedPorts) {
+    public boolean isNone() {
+        return this.repository.equals("<none>") && this.tag.equals("<none>");
+    }
+
+    public String repositoryWithTag() {
+        return repository + ":" + tag;
+    }
+}
