+++ docker/src/main/java/edu/kit/kastel/informalin/framework/docker/DockerPortBind.java
@@ -1,0 +1,8 @@
+/* Licensed under MIT 2022. */
+package edu.kit.kastel.informalin.framework.docker;
+
+public record DockerPortBind(int hostPort, int containerPort, boolean wildcard) {
+    public boolean valid() {
+        return hostPort > 0 && containerPort > 0;
+    }
+}
